--- a/digital_land/commands.py
+++ b/digital_land/commands.py
@@ -249,8 +249,11 @@ def dataset_create(
         specification_dir=None,  # TBD: package should use this specification object
     )
     package.create()
+    package.connect(optimised=True)
+
     for path in input_paths:
         package.load_transformed(path)
+
     package.load_entities()
 
     old_entity_path = os.path.join(pipeline.path, "old-entity.csv")
@@ -265,6 +268,8 @@ def dataset_create(
         logging.warning("No directory for this dataset in the provided issue_directory")
 
     package.add_counts()
+
+    package.disconnect()
 
 
 def dataset_dump(input_path, output_path):