The purpose of the change is to speed up the creation of dataset sqlite3 files.
The analysis summary follows, which I believe justifies the change.

### Performance Metrics of Optimization Changes
There are three datasets in the listed-building collection:

- listed-building
- listed-building-grade
- listed-building-outline

Each dataset has to process it's corresponding issue log file.

Total size of Issue Logs for each dataset

- listed-building 138,476 KB
- listed-building-grade 1 KB
- listed-building-outline 1,080 KB

Listed building clearly has the most work to do when generating the issue-log table.

The listed-building dataset has the largest transform file, containing 3412530 records. This file was 'cut' into different sizes to get the timing set below.

| Run |       # Records Processed | Current Solution (s) | New Solution (s) |     % | 
|---:|--------------------------:|---------------------:|-----------------:|:------:|
|  1  |                      1000 |                23.80 |             9.57 | 40.21 |
|  2  |                     64000 |                30.39 |            13.81 | 45.44 |
|  3  |                    512000 |                87.02 |            54.45 | 62.57 |
|  4  |                   1024000 |               157.59 |           113.61 | 72.09 |
|  5  |                   2048000 |               312.41 |           235.45 | 75.37 |
|  6  |                   3412530 |               518.65 |           384.26 | 74.09 |

Table shows the number of seconds taken to process a given number of listed-building transform file records (+ issue-log records from above).

### Summary
It’s clear the savings made from optimization diminish with increased transform file size.
The savings reduce to about 25% for very large datasets, but can be as much as 60% for smaller datasets.

It’s worth noting that, although the savings from large datasets like listed-building is small, the savings from processing all collections should be significant enough to see an overall benefit from these changes.