@morgan-sl 2023-10-20
Revision II attempt to address many of the concerns raised above.